Move MetadataVersionConfigValidator and related test code to metadata module
